+++ gatsby/src/config.ts
@@ -1,18 +1,4 @@
-export type Config = {
-    version: string
-    siteUrl: string
-    siteTitle: string
-    siteDescription: string
-    siteKeywords: string
-    siteDisplayname: string
-    siteEmail: string
-    siteIcon: string
-    siteImage: string 
-    siteAvatar: string    
-    siteRepo: string
-    siteTheme: string
-    isomorphic: any
-}
+import {Config} from "./goldlabel/types"
 
 export const glConfig: Config = {
     version: "4.3.3",
