+++ gatsby/src/goldlabel/types.d.ts
@@ -1,0 +1,15 @@
+export type Config = {
+    version: string
+    siteUrl: string
+    siteTitle: string
+    siteDescription: string
+    siteKeywords: string
+    siteDisplayname: string
+    siteEmail: string
+    siteIcon: string
+    siteImage: string 
+    siteAvatar: string    
+    siteRepo: string
+    siteTheme: string
+    isomorphic: any
+}
